requirements.txt
openai==0.27.8
slack_bolt==1.18.0
google-cloud-firestore==2.11.1
sentry-sdk==1.27.1
tiktoken==0.7.0
joblib~=1.3.2
flask~=2.3.3
replicate==0.23.1
requests~=2.31.0
Pillow~=10.0.0
firebase-admin~=6.2.0
holidays~=0.36